unexpected long retry behavior change
The long version was returning the original function with the decorator already attached, so the callstack would look like:

valreturn_imgx3
valreturn_long
valreturn_img x3
valreturn_long
...

So I just copy pasted the original function over instead of returning it. 
There is probably a way to override the decorator arguments of a function that already has a decorator but I will look into it later.